account_move_template: to 10.0 (#511)
* [MIG] account_move_template: to 10.0

All working, removed feature of transitory account, I think that feature
is redundant because it can be solved with python code lines and/or a
proper server action per client, beside it was not properly documented.

- All code refactored more new api oriented.
- Tested with enterprise and all views fixed.
- Removed all licences on files and moved as usual to where they belong.

* [FIX] removed tax from view.

The taxes should be triggered from the account itself and not wired into
the template.

* [FIX] account_move_template: Activities in just for saas-x comming on
v100, for now I left it commented to simply uncomment it on v11

* [FIX] account_move_template: Cleanup translations, not useful after
refactor

* [REF] account_move_template: Removing all reference to taxes.

* [IMP] account_move_template: Translating

* [FIX] account_move_template: Fixing lints.