[FIX] account_journal_restrict_mode: Make compatible with other modules
When this module is installed, it prevents the cancellation of posted journals. However, in many cases, this action is necessary—for example, in the account_asset_management module.

With this commit, the default constraint is applied only in this module's tests or when Odoo is not running in test mode.

This approach provides an alternative to avoid marking the module as rebellious.