[FIX] account_chart_update: fix HTML fields translation
Before this patch, when a CoA had multiple translations and those languages were active in the database, all translations would get updated to the one that was missing.

For example, if the English translation was missing, then all languages would get the English terms. Then, if the Italian translation was missing, in the next update, all languages would get the Italian terms.

Now every language will get their appropriate translation if defined, or the default one otherwise. And it is now consistent.

The code is simpler and more complex at the same time. It is simpler because we now leverage upstream `_load_translations()` method, which is more performant and should handle HTML translations just fine. But it is more complex too, because I have to include a workaround for https://github.com/odoo/odoo/pull/234333 (a bug in how upstream handles HTML translations). And since they don't expose many hooks in that method, I have to use `unittest.mock.patch` to interfere in the desired spot. Hopefully, they'll merge that PR some day and we will just have to call `_load_translations()`.

@moduon MT-12331